requirements.txt
# Python dependencies for WHY2025 Badge Development
# Used by the devcontainer's Python virtual environment

# Core development
jupyter>=1.0.0
jupyterlab>=3.0.0
ipywidgets>=7.6.0

# Data processing and ML
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scipy>=1.7.0
scikit-learn>=1.0.0
tensorflow-cpu>=2.8.0

# ESP32 and embedded development
pyserial>=3.5
esptool>=4.0.0
platformio>=6.0.0

# Edge ML tools
edge-impulse-cli>=1.0.0

# Development tools
pytest>=6.0.0
black>=22.0.0
flake8>=4.0.0
pre-commit>=2.15.0

# Data serialization
PyYAML>=6.0.0
h5py>=3.6.0

# Utilities
tqdm>=4.62.0
requests>=2.26.0